move Diffusion model to another file
