+++ main.py
@@ -2,16 +2,16 @@ import sys
 import numpy as np
 from tqdm import tqdm
 from pathlib import Path
-from pympler.tracker import SummaryTracker
+from apps.diffusion import Unet, Diffusion
 
 sys.path.append('./python')
 
 import needle as ndl
-import needle.nn as nn
 from needle import array_api, Tensor, NDArray
 
 DATA_DIR = Path('./data/landscapes/')
 IMAGE_FILES = sorted(str(x) for x in list(DATA_DIR.rglob('*.jpg')))
+
 
 if __name__ == '__main__':
     transforms = [
@@ -23,7 +23,7 @@ if __name__ == '__main__':
     )
 
     device = ndl.cpu()
-    model = nn.Unet(device=device)
+    model = Unet(device=device)
     optimizer = ndl.optim.Adam(model.parameters(), lr=1e-3)
 
     epochs = 1
@@ -31,14 +31,10 @@ if __name__ == '__main__':
     dataloader = ndl.data.DataLoader(dataset, batch_size)
 
     timesteps = 50
-    module = nn.Diffusion(model, timesteps, loss_type="l2",
+    module = Diffusion(model, timesteps, loss_type="l2",
                           device=device)
 
-    tracker = SummaryTracker()
-
-    result = module.sample(64, 4, 3)
-
-    for epoch in range(0):
+    for epoch in range(epochs):
         for step, batch in enumerate(tqdm(dataloader)):
             # Algorithm 1 line 3: sample t uniformally for every example in the batch
             optimizer.reset_grad()
@@ -54,3 +50,5 @@ if __name__ == '__main__':
 
             loss.backward()
             optimizer.step()
+
+    result = module.sample(64, 4, 3)
